[change] change models to a dir
