+++ article_rec_db/models/__init__.py
@@ -1,0 +1,4 @@
+__all__ = ["Article", "Page"]
+
+from .article import Article
+from .page import Page
